整理当前 objects 中的节点（一般表达式）
