+++ metasequoia_sql/core/__init__.py
@@ -1,0 +1,1 @@
+from metasequoia_sql.core.data_source import DataSource
