--- metasequoia_sql/objects/__init__.py
@@ -1,10 +1,0 @@
-from metasequoia_sql.objects.data_source import DataSource
-from metasequoia_sql.objects.sql_compare_operator import SQLCompareOperator
-from metasequoia_sql.objects.sql_compute_operator import SQLComputeOperator
-from metasequoia_sql.objects.sql_general_expression import SQLGeneralExpression
-from metasequoia_sql.objects.sql_insert_type import SQLInsertType
-from metasequoia_sql.objects.sql_join_type import SQLJoinType
-from metasequoia_sql.objects.sql_literal_expression import SQLLiteralExpression
-from metasequoia_sql.objects.sql_logical_operator import SQLLogicalOperator
-from metasequoia_sql.objects.sql_order_type import SQLOrderType
-from metasequoia_sql.objects.sql_union_type import SQLUnionType
